+++ UnitTests/CardPaymentsTests/CardClient_Tests.swift
@@ -5,7 +5,6 @@ import AuthenticationServices
 @testable import CardPayments
 @testable import TestShared
 
-// swiftlint:disable type_body_length
 class CardClient_Tests: XCTestCase {
 
     // MARK: - Helper Properties
@@ -143,6 +142,8 @@ class CardClient_Tests: XCTestCase {
 
         let mockCardDelegate = MockCardDelegate(success: {_, result -> Void in
             XCTAssertEqual(result.orderID, "testOrderId")
+            XCTAssertEqual(result.status, "APPROVED")
+            XCTAssertFalse(result.didAttemptThreeDSecureAuthentication)
             expectation.fulfill()
         }, error: { _, _ in
             XCTFail("Invoked error() callback. Should invoke success().")
@@ -214,7 +215,8 @@ class CardClient_Tests: XCTestCase {
         let mockCardDelegate = MockCardDelegate(
             success: {_, result in
                 XCTAssertEqual(result.orderID, "testOrderId")
-                XCTAssertEqual(result.liabilityShift, "POSSIBLE")
+                XCTAssertNil(result.status)
+                XCTAssertTrue(result.didAttemptThreeDSecureAuthentication)
                 expectation.fulfill()
             },
             error: { _, error in
@@ -291,93 +293,6 @@ class CardClient_Tests: XCTestCase {
             },
             threeDSWillLaunch: { _ in XCTAssert(true) },
             threeDSLaunched: { _ in XCTAssert(true) })
-
-        sut.delegate = mockCardDelegate
-        sut.approveOrder(request: cardRequest)
-
-        waitForExpectations(timeout: 10)
-    }
-    
-    func testApproveOrder_withThreeDSecure_returns3DSVerificationError() {
-        mockCheckoutOrdersAPI.stubConfirmResponse = FakeConfirmPaymentResponse.withValid3DSURL
-
-        mockWebAuthSession.cannedResponseURL = .init(string: "sdk.ios.paypal://card/success?error=error&error_description=error&liability_shift=NO")
-
-        let expectation = expectation(description: "approveOrder() completed")
-
-        let mockCardDelegate = MockCardDelegate(
-            success: {_, _ in
-                XCTFail("Invoked success() callback. Should invoke error().")
-                expectation.fulfill()
-            },
-            error: { _, error in
-                XCTAssertEqual(error.domain, CardClientError.domain)
-                XCTAssertEqual(error.code, CardClientError.Code.threeDSVerificationError.rawValue)
-                XCTAssertEqual(error.errorDescription, "3DS Verification is returning an error.")
-                expectation.fulfill()
-            },
-            cancel: { _ in
-                XCTFail("Invoked cancel() callback. Should invoke error().")
-                expectation.fulfill()
-            },
-            threeDSWillLaunch: { _ in XCTAssert(true) },
-            threeDSLaunched: { _ in XCTAssert(true) })
-
-        sut.delegate = mockCardDelegate
-        sut.approveOrder(request: cardRequest)
-
-        waitForExpectations(timeout: 10)
-    }
-    
-    func testApproveOrder_withThreeDSecure_returnsMissingDeeplinkError() {
-        mockCheckoutOrdersAPI.stubConfirmResponse = FakeConfirmPaymentResponse.withValid3DSURL
-        
-        mockWebAuthSession.cannedResponseURL = nil
-        
-        let expectation = expectation(description: "approveOrder() completed")
-
-        let mockCardDelegate = MockCardDelegate(
-            success: {_, _ in
-                XCTFail("Invoked success() callback. Should invoke error().")
-                expectation.fulfill()
-            },
-            error: { _, error in
-                XCTAssertEqual(error.domain, CardClientError.domain)
-                XCTAssertEqual(error.code, CardClientError.Code.missingDeeplinkURLError.rawValue)
-                XCTAssertEqual(error.errorDescription, "Missing deeplink URL from 3DS.")
-                expectation.fulfill()
-            },
-            cancel: { _ in XCTFail("Invoked cancel() callback. Should invoke success().") },
-            threeDSWillLaunch: { _ -> Void in XCTAssert(true) },
-            threeDSLaunched: { _ -> Void in XCTAssert(true) })
-
-        sut.delegate = mockCardDelegate
-        sut.approveOrder(request: cardRequest)
-
-        waitForExpectations(timeout: 10)
-    }
-    
-    func testApproveOrder_withThreeDSecure_returnsMalformedDeeplinkError() {
-        mockCheckoutOrdersAPI.stubConfirmResponse = FakeConfirmPaymentResponse.withValid3DSURL
-        
-        mockWebAuthSession.cannedResponseURL = .init(string: "https://fakeURL")
-        
-        let expectation = expectation(description: "approveOrder() completed")
-
-        let mockCardDelegate = MockCardDelegate(
-            success: {_, _ in
-                XCTFail("Invoked success() callback. Should invoke error().")
-                expectation.fulfill()
-            },
-            error: { _, error in
-                XCTAssertEqual(error.domain, CardClientError.domain)
-                XCTAssertEqual(error.code, CardClientError.Code.malformedDeeplinkURLError.rawValue)
-                XCTAssertEqual(error.errorDescription, "Malformed deeplink URL.")
-                expectation.fulfill()
-            },
-            cancel: { _ in XCTFail("Invoked cancel() callback. Should invoke success().") },
-            threeDSWillLaunch: { _ -> Void in XCTAssert(true) },
-            threeDSLaunched: { _ -> Void in XCTAssert(true) })
 
         sut.delegate = mockCardDelegate
         sut.approveOrder(request: cardRequest)